Modify f32 -> f64, rework parts to make more concise

+++ src/conf_sampling/peptide.rs
@@ -3,12 +3,12 @@ use pyo3::{pyclass, pymethods};
 
 
 /// the `phi-psi` dihedrals, which are the peptide backbone dihedrals in proteins
-/// public `phi` field : Vec<f32>
-/// public `psi` field : Vec<f32>
+/// public `phi` field : Vec<f64>
+/// public `psi` field : Vec<f64>
 #[pyclass(get_all)]
 pub struct Peptide {
-    phi : Vec<f32>,
-    psi : Vec<f32>,
+    phi : Vec<f64>,
+    psi : Vec<f64>,
 }
 
 #[pymethods]
@@ -24,13 +24,13 @@ impl Peptide {
         let mut phi = Vec::with_capacity(amount);
         let mut psi = Vec::with_capacity(amount);
         
-        let mut xi : f32;
-        let mut yi : f32;
+        let mut xi : f64;
+        let mut yi : f64;
         for i in 0..amount as usize {
 
             // For every x value, return all y values
-            xi = (i as f32 / num as f32).floor(); // floor, to return x axis value
-            yi = i as f32 % num as f32; // return with modulo, to return y axis value
+            xi = (i as f64 / num as f64).floor(); // floor, to return x axis value
+            yi = i as f64 % num as f64; // return with modulo, to return y axis value
 
             // fill out the array
             phi.push(axes.x[xi as usize]); 
@@ -47,13 +47,13 @@ impl Peptide {
 /// The axes to iterate over for peptide-like molecules : 
 /// Its extent is : [0 , 2pi] (rad)
 /// Its extent is : [0 , 360] (degrees)
-/// public `x` field : Vec<f32>
-/// public `y` field : Vec<f32>
+/// public `x` field : Vec<f64>
+/// public `y` field : Vec<f64>
 /// Can remain a private struct, as this only is required to build the Peptide struct
 //#[pyclass(get_all)]
 pub struct PeptideAxes {
-    x : Vec<f32>,
-    y : Vec<f32>,
+    x : Vec<f64>,
+    y : Vec<f64>,
 
 }
